pyproject.toml
[project]
name = "a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk==0.3.7",
    "adk==0.0.5",
    "aiohttp>=3.13.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asyncio>=4.0.0",
    "attrs==25.3.0",
    "authlib==1.6.0",
    "automat==25.4.16",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==2.0.0",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "constantly==23.10.4",
    "cryptography==46.0.2",
    "cssselect==1.3.0",
    "defusedxml==0.7.1",
    "docstring-parser==0.16",
    "exceptiongroup==1.3.0",
    "fastapi==0.115.12",
    "fastmcp>=2.8.0",
    "filelock==3.19.1",
    "google-adk==1.2.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.0",
    "google-api-python-client==2.171.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-cloud-aiplatform==1.96.0",
    "google-cloud-appengine-logging==1.6.1",
    "google-cloud-audit-log==0.3.2",
    "google-cloud-bigquery==3.34.0",
    "google-cloud-core==2.4.3",
    "google-cloud-logging==3.12.1",
    "google-cloud-resource-manager==1.14.2",
    "google-cloud-secret-manager==2.24.0",
    "google-cloud-speech==2.32.0",
    "google-cloud-storage==2.19.0",
    "google-cloud-trace==1.16.1",
    "google-crc32c==1.7.1",
    "google-genai==1.19.0",
    "google-generativeai==0.8.5",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "graphviz==0.20.3",
    "greenlet==3.2.3",
    "grpc-google-iam-v1==0.14.2",
    "grpcio==1.73.0",
    "grpcio-reflection==1.71.0",
    "grpcio-status==1.71.0",
    "grpcio-tools==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "hyperlink==21.0.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "incremental==24.7.2",
    "itemadapter==0.12.2",
    "itemloaders==1.3.2",
    "jmespath==1.0.1",
    "lxml==6.0.2",
    "mcp[cli]==1.9.3",
    "nest-asyncio==1.6.0",
    "numpy==2.3.0",
    "ollama==0.6.0",
    "opentelemetry-api==1.34.1",
    "opentelemetry-exporter-gcp-trace==1.9.0",
    "opentelemetry-resourcedetector-gcp==1.9.0a0",
    "opentelemetry-sdk==1.34.1",
    "opentelemetry-semantic-conventions==0.55b1",
    "packaging==25.0",
    "parsel==1.10.0",
    "protego==0.5.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pydispatcher==2.0.7",
    "pyopenssl==25.3.0",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "queuelib==1.8.0",
    "requests==2.32.4",
    "requests-file==2.1.0",
    "rsa==4.9.1",
    "scrapy==2.13.3",
    "service-identity==24.2.0",
    "setuptools==80.9.0",
    "shapely==2.1.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sse-starlette==2.3.6",
    "starlette==0.46.2",
    "tabulate==0.9.0",
    "tldextract==5.3.0",
    "tqdm==4.67.1",
    "twisted==25.5.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "uritemplate==4.2.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.3",
    "w3lib==2.3.1",
    "websockets==15.0.1",
    "zipp==3.23.0",
    "zope-interface==8.0.1",
]
